#14 refactor: revise code base struct to support async scanning
