+++ Cargo.toml
@@ -11,3 +11,4 @@ phf = { version = "0.13.1", features = ["macros"] }
 pinger = "2.1.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+tokio = { version = "1", features = ["full"] }
